MNT: more files

+++ use_dotfiles.sh
@@ -9,9 +9,12 @@ if [ ! -d dotfiles ]; then
     exit 1;
 fi
 
-ln -s dotfiles/bashrc .bashrc
-ln -s dotfiles/bash_aliases .bash_aliases
-ln -s dotfiles/bash_profile .bash_profile
+[ ! -f .bashrc ] && ln -s dotfiles/bashrc .bashrc
+[ ! -f .bash_aliases ] && ln -s dotfiles/bash_aliases .bash_aliases
+[ ! -f .bash_profile ] && ln -s dotfiles/bash_profile .bash_profile
+[ ! -f .vimrc ] && ln -s dotfiles/vimrc .vimrc
+[ ! -f .condarc ] && ln -s dotfiles/condarc .condarc
+[ ! -f .gitconfig ] && ln -s dotfiles/gitconfig .gitconfig
 
 mkdir -p .ssh
-ln -s dotfiles/ssh/config .ssh/config
+[ ! -f .ssh/config ] && ln -s dotfiles/ssh/config .ssh/config
